[#16205] Extract testing container to another dep
