+++ server/testdriver/pom.xml
@@ -16,5 +16,6 @@
       <module>core</module>
       <module>junit4</module>
       <module>junit5</module>
+      <module>testcontainers</module>
    </modules>
 </project>
