+++ server/testdriver/testcontainers/src/license/licenses.xml
@@ -1,0 +1,11 @@
+<licenseSummary>
+   <knownLicenses>
+      <license name="LGPL-2.1-or-later" url="https://www.gnu.org/licenses/old-licenses/lgpl-2.1"/>
+      <license name="BSD License 3" url="http://opensource.org/licenses/BSD-3-Clause"/>
+      <license name="BSD-2-Clause" url="https://opensource.org/licenses/BSD-2-Clause"/>
+      <license name="MIT" url="http://opensource.org/licenses/MIT"/>
+      <license name="the mit license" url="http://opensource.org/licenses/MIT"/>
+   </knownLicenses>
+   <!-- required or it fails to parse -->
+   <dependencies/>
+</licenseSummary>
