+++ server/tests/src/test/java/org/infinispan/server/functional/VersionsInteropTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.infinispan.client.hotrod.DefaultTemplate;
 import org.infinispan.client.hotrod.RemoteCache;
 import org.infinispan.client.hotrod.RemoteCacheManager;
-import org.infinispan.server.test.core.InfinispanContainer;
+import org.infinispan.testcontainers.InfinispanContainer;
 import org.junit.jupiter.api.Test;
 
 public class VersionsInteropTest {
@@ -16,7 +16,7 @@ public class VersionsInteropTest {
       // We start a server with an older version than the client, they should negotiate the protocol correctly.
       try (InfinispanContainer container = new InfinispanContainer("quay.io/infinispan/server:13.0")) {
          container.start();
-         try (RemoteCacheManager cacheManager = container.getRemoteCacheManager()) {
+         try (RemoteCacheManager cacheManager = new RemoteCacheManager(container.getConnectionURI())) {
             RemoteCache<Object, Object> testCache = cacheManager.administration().getOrCreateCache("test", DefaultTemplate.DIST_SYNC);
             testCache.put("key", "value");
             assertEquals("value", testCache.get("key"));
