+++ server/tests/src/test/java/org/infinispan/server/resilience/GracefulShutdownRestartIT.java
@@ -17,10 +17,10 @@ import org.infinispan.commons.test.Eventually;
 import org.infinispan.configuration.cache.CacheMode;
 import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.infinispan.server.test.core.ContainerInfinispanServerDriver;
-import org.infinispan.server.test.core.CountdownLatchLoggingConsumer;
 import org.infinispan.server.test.core.ServerRunMode;
 import org.infinispan.server.test.junit5.InfinispanServerExtension;
 import org.infinispan.server.test.junit5.InfinispanServerExtensionBuilder;
+import org.infinispan.testcontainers.CountdownLatchLoggingConsumer;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
